WIP: getting closer. Add i_flapping info to all proposed projs @ B10.
The placement of this flapping stuff is suboptimal.  I'd prefer it to be
at A20 where all of the real projection calculation is done.  But this
is still a prototype ... I'm still trying to figure out where & how is
the best place to react.  Moving the calculation from B-something to
A-something can come later.