Good news: all_hosed list agreement in all simulated cases & is correct (mostly!)
Occasionally, all_hosed will include a node from the previous
simulation round that isn't included in the current round.
I believe that this is expected, since machi_chain_manager1_test
is not trying to "reset" state to no-partitions before switching
to the next partition scenario.

Please note that the #ch_mgr.flap_limit constant is a *hack*
and needs a better implementation!