WIP: Wow, it looks like it's working, spiffy!
In order for this solution to work, we need to add another
requirement to wedging: a FLU will not automatically wedge itself
when a new projection P is written to its public store *if* P
contains a nested "inner" projection.  If P has an inner projection,
then the epoch # of the inner projection is used for communicating
with clients and wrt to wedging.

One known problem with the simulator is that during flap damping,
the inner projection will not advance to 'repair finished', due to
some missing code in check_latest_private_projections(); see the
comment in that function for more.