End of the line for this branch, part II (more).
My comments on the previous commit, 56d03a9b, still apply.
The funny/hilarious/sad thing is that this commit appears to be working,
nearly all the time?  I haven't given it a hundred hours of CPU time,
but in all of the simple cases that have tricked earlier code into doing
something stupid, this version appears to work well enough, wow.

Sample invocations:

    machi_chain_manager1_converge_demo:t(3, [{private_write_verbose,true}, {consistency_mode, cp_mode}, {witnesses, [a]}]).
    machi_chain_manager1_converge_demo:t(4, [{private_write_verbose,true}, {consistency_mode, cp_mode}, {witnesses, [a,b]}]).
    machi_chain_manager1_converge_demo:t(5, [{private_write_verbose,true}, {consistency_mode, cp_mode}, {witnesses, [a,b,c]}]).

{sigh}  So, it's good to know that all of this work has not been
completely in vain.  There are some small bugs, one minor and one
bigger-than-minor.

1. There's at least one corner case (see one of the new comments in
make_partition_list()) where the none projection is written repeatedly
using the same epoch number (which isn't a true safety violation
but it's a waste of CPU cycles).

2. There's a larger problem where the LeastCompatibleInner check in
react_to_env_A30() is returning false when it shouldn't.  If we're
flapping, and someone else suggests upi=[a,c] and we ought be suggesting
upi=[a,c] also, but because A29 forced our current projection P_current
backward to a much earlier projection (but zerf-verified!) upi=[b,c],
then we don't use P_latest ... and the result is a never-ending
escalation of epoch number with the exact same UPI+repairing lists.
{sigh}

Also, there's a suboptimal case where we have a upi=[a,c] and a is a
witness and b is down (chain len=3).  If a fails and b returns to
service, the current code does not allow b to serve as a witness-like
server to allow the UPI to return to a majority quorum.  Instead, the
code requires a to return to service, then upi=[a,c],repairing=[b],
and then repair of b must finish successfully before the hassle can be
permanently resolved.